requirements.txt
langchain
streamlit
langchain-community
langchain-huggingface
transformers
python-dotenv
pandas
numpy
matplotlib
seaborn
kagglehub
faiss-cpu
tqdm
bs4
selenium
webdriver-manager
datasets
accelerate>=0.26.0
transformers[torch]
ipywidgets

